# src/WebSdk/openapi/engine.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Micser.Engine | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/devices": {
      "get": {
        "tags": [
          "Devices"
        ],
        "operationId": "GetDevices",
        "parameters": [
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DeviceDirection"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AudioDeviceInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/modules": {
      "get": {
        "tags": [
          "Modules"
        ],
        "operationId": "GetModules",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Modules"
        ],
        "operationId": "CreateModule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/modules/{id}": {
      "get": {
        "tags": [
          "Modules"
        ],
        "operationId": "GetModule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Modules"
        ],
        "operationId": "UpdateModule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Modules"
        ],
        "operationId": "DeleteModule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/connections": {
      "get": {
        "tags": [
          "Connections"
        ],
        "operationId": "GetConnections",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Connections"
        ],
        "operationId": "CreateConnection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/connections/{id}": {
      "delete": {
        "tags": [
          "Connections"
        ],
        "operationId": "DeleteConnection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/engine/settings": {
      "put": {
        "tags": [
          "Engine"
        ],
        "operationId": "UpdateEngineSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngineSettingsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngineStatusDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/plugins": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "operationId": "GetPlugins",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PluginDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Plugins"
        ],
        "description": "Stages a plugin package (a zip with plugin.json at its root) for installation when the engine restarts.",
        "operationId": "InstallPlugin",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "package"
                ],
                "type": "object",
                "properties": {
                  "package": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/plugins/{id}": {
      "delete": {
        "tags": [
          "Plugins"
        ],
        "description": "Stages a user plugin for removal when the engine restarts, or cancels its staged installation.",
        "operationId": "RemovePlugin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/preferences": {
      "get": {
        "tags": [
          "Preferences"
        ],
        "operationId": "GetPreferences",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UiPreferencesDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Preferences"
        ],
        "operationId": "UpdatePreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UiPreferencesDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UiPreferencesDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/subgraphs": {
      "get": {
        "tags": [
          "Subgraphs"
        ],
        "operationId": "GetSubgraphs",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubgraphDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Subgraphs"
        ],
        "description": "Creates a subgraph and moves the modules into it.",
        "operationId": "CreateSubgraph",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubgraphRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubgraphDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/subgraphs/{id}": {
      "put": {
        "tags": [
          "Subgraphs"
        ],
        "operationId": "UpdateSubgraph",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubgraphDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubgraphDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Subgraphs"
        ],
        "description": "Removes a subgraph; its modules stay in the graph.",
        "operationId": "DeleteSubgraph",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AudioDeviceInfo": {
        "required": [
          "id",
          "name",
          "description",
          "adapterName",
          "direction",
          "isActive",
          "layout",
          "sampleRate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "adapterName": {
            "type": [
              "null",
              "string"
            ]
          },
          "direction": {
            "$ref": "#/components/schemas/DeviceDirection"
          },
          "isActive": {
            "type": "boolean"
          },
          "layout": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChannelLayout"
              }
            ]
          },
          "sampleRate": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
      "ChannelLayout": {
        "type": "object",
        "properties": {
          "channelCount": {
            "type": "integer",
            "format": "int32"
          },
          "speakers": {
            "$ref": "#/components/schemas/Speakers"
          },
          "hasSpeakerPositions": {
            "type": "boolean"
          }
        }
      },
...
      "ConnectionDto": {
        "required": [
          "id",
          "sourceModuleId",
          "sourcePort",
          "targetModuleId",
          "targetPort"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sourceModuleId": {
            "type": "string",
            "format": "uuid"
          },
          "sourcePort": {
            "type": "string"
          },
          "targetModuleId": {
            "type": "string",
            "format": "uuid"
          },
          "targetPort": {
            "type": "string"
          }
        }
      },
      "CreateConnectionRequest": {
        "required": [
          "sourceModuleId",
          "sourcePort",
          "targetModuleId",
          "targetPort"
        ],
        "type": "object",
        "properties": {
          "sourceModuleId": {
            "type": "string",
            "format": "uuid"
          },
          "sourcePort": {
            "type": "string"
          },
          "targetModuleId": {
            "type": "string",
            "format": "uuid"
          },
          "targetPort": {
            "type": "string"
          }
        }
      },
      "CreateModuleRequest": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "position": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ModulePosition"
              }
            ]
          },
          "subgraphId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "CreateSubgraphRequest": {
        "required": [
          "name",
          "position",
          "size",
          "moduleIds",
          "color"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "position": {
            "$ref": "#/components/schemas/ModulePosition"
          },
          "size": {
            "$ref": "#/components/schemas/SubgraphSize"
          },
          "moduleIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "color": {
            "$ref": "#/components/schemas/SubgraphColor"
          }
        }
      },
      "DeviceDirection": {
        "enum": [
          "Input",
          "Output"
        ]
      },
...
      "EngineSettingsDto": {
        "required": [
          "sampleRate",
          "frameCount"
        ],
        "type": "object",
        "properties": {
          "sampleRate": {
            "type": "integer",
            "format": "int32",
            "default": 48000
          },
          "frameCount": {
            "type": "integer",
            "format": "int32",
            "default": 240
          }
        }
      },
      "EngineStatusDto": {
        "required": [
          "isRunning",
          "settings",
          "blocks",
          "lateBlocks",
          "maxProcessingMilliseconds"
        ],
        "type": "object",
        "properties": {
          "isRunning": {
            "type": "boolean"
          },
          "settings": {
            "$ref": "#/components/schemas/EngineSettingsDto"
          },
          "blocks": {
            "type": "integer",
            "format": "int64"
          },
          "lateBlocks": {
            "type": "integer",
            "format": "int64"
          },
          "maxProcessingMilliseconds": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "ModuleDto": {
        "required": [
          "type"
        ],
        "type": "object",
        "anyOf": [
          {
            "$ref": "#/components/schemas/ModuleDtoDeviceInputModule"
          },
          {
            "$ref": "#/components/schemas/ModuleDtoLoopbackInputModule"
          },
          {
            "$ref": "#/components/schemas/ModuleDtoDeviceOutputModule"
          },
          {
            "$ref": "#/components/schemas/ModuleDtoGainModule"
          },
          {
            "$ref": "#/components/schemas/ModuleDtoCompressorModule"
          },
          {
            "$ref": "#/components/schemas/ModuleDtoEqualizerModule"
          },
          {
            "$ref": "#/components/schemas/ModuleDtoPitchModule"
          },
          {
            "$ref": "#/components/schemas/ModuleDtoSpectrumModule"
          }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "DeviceInput": "#/components/schemas/ModuleDtoDeviceInputModule",
            "LoopbackInput": "#/components/schemas/ModuleDtoLoopbackInputModule",
            "DeviceOutput": "#/components/schemas/ModuleDtoDeviceOutputModule",
            "Gain": "#/components/schemas/ModuleDtoGainModule",
            "Compressor": "#/components/schemas/ModuleDtoCompressorModule",
            "Equalizer": "#/components/schemas/ModuleDtoEqualizerModule",
            "Pitch": "#/components/schemas/ModuleDtoPitchModule",
            "Spectrum": "#/components/schemas/ModuleDtoSpectrumModule"
          }
        }
      },
...
      "ModuleDtoDeviceInputModule": {
        "required": [
          "state",
          "id",
          "isBypassed",
          "isCollapsed",
          "isMuted",
          "useSystemVolume",
          "volume"
        ],
        "properties": {
          "type": {
            "enum": [
              "DeviceInput"
            ],
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/DeviceInputState"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isBypassed": {
            "type": "boolean"
          },
          "isCollapsed": {
            "type": "boolean"
          },
          "isMuted": {
            "type": "boolean"
          },
          "name": {
            "maxLength": 100,
            "type": [
              "null",
              "string"
            ]
          },
          "position": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ModulePosition"
              }
            ]
          },
          "subgraphId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "useSystemVolume": {
            "type": "boolean"
          },
          "volume": {
            "maximum": 1,
            "minimum": 0,
            "type": "number",
            "format": "float"
          }
        }
      },
      "ModuleDtoDeviceOutputModule": {
        "required": [
          "state",
          "id",
          "isBypassed",
          "isCollapsed",
          "isMuted",
          "useSystemVolume",
          "volume"
        ],
        "properties": {
          "type": {
            "enum": [
              "DeviceOutput"
            ],
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/DeviceOutputState"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isBypassed": {
            "type": "boolean"
          },
          "isCollapsed": {
            "type": "boolean"
          },
          "isMuted": {
            "type": "boolean"
          },
          "name": {
            "maxLength": 100,
            "type": [
              "null",
              "string"
            ]
          },
          "position": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ModulePosition"
              }
            ]
          },
          "subgraphId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "useSystemVolume": {
            "type": "boolean"
          },
          "volume": {
            "maximum": 1,
            "minimum": 0,
            "type": "number",
            "format": "float"
          }
        }
      },
...
      "ModuleDtoLoopbackInputModule": {
        "required": [
          "state",
          "id",
          "isBypassed",
          "isCollapsed",
          "isMuted",
          "useSystemVolume",
          "volume"
        ],
        "properties": {
          "type": {
            "enum": [
              "LoopbackInput"
            ],
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/LoopbackInputState"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isBypassed": {
            "type": "boolean"
          },
          "isCollapsed": {
            "type": "boolean"
          },
          "isMuted": {
            "type": "boolean"
          },
          "name": {
            "maxLength": 100,
            "type": [
              "null",
              "string"
            ]
          },
          "position": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ModulePosition"
              }
            ]
          },
          "subgraphId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "useSystemVolume": {
            "type": "boolean"
          },
          "volume": {
            "maximum": 1,
            "minimum": 0,
            "type": "number",
            "format": "float"
          }
        }
      },
...
      "PluginChange": {
        "enum": [
          "None",
          "Install",
          "Remove"
        ]
      },
      "PluginDto": {
        "required": [
          "id",
          "name",
          "version",
          "isBuiltIn",
          "isLoaded",
          "error",
          "webUrl",
          "pendingChange"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "version": {
            "type": [
              "null",
              "string"
            ]
          },
          "isBuiltIn": {
            "type": "boolean"
          },
          "isLoaded": {
            "type": "boolean"
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          },
          "webUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "pendingChange": {
            "$ref": "#/components/schemas/PluginChange"
          }
        }
      },
      "Speakers": {
        "type": "string"
      },
...
      "SubgraphDto": {
        "required": [
          "id",
          "name",
          "position",
          "size",
          "color",
          "isCollapsed",
          "isMuted",
          "isBypassed"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "position": {
            "$ref": "#/components/schemas/ModulePosition"
          },
          "size": {
            "$ref": "#/components/schemas/SubgraphSize"
          },
          "color": {
            "$ref": "#/components/schemas/SubgraphColor"
          },
          "isCollapsed": {
            "type": "boolean",
            "default": false
          },
          "isMuted": {
            "type": "boolean",
            "default": false
          },
          "isBypassed": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "UiPreferencesDto": {
        "required": [
          "showStreamStatistics",
          "snapToGrid"
        ],
        "type": "object",
        "properties": {
          "showStreamStatistics": {
            "type": "boolean",
            "default": false
          },
          "snapToGrid": {
            "type": "boolean",
            "default": true
          }
        }
      }
    }
  },
...
}